// src/app/modules/shared/modules/git/models/git.commands.enum.ts
export enum GitCommandsEnum {
  ADD = 'add',
  BRANCH = 'branch',
  CHECKOUT = 'checkout',
  COMMIT = 'commit',
  CHERRY_PICK = 'cherry-pick',
  CLEAN = 'clean',
  CLONE = 'clone',
  CONFIG = 'config',
  DIFF = 'diff',
  FETCH = 'fetch',
  INIT = 'init',
  LOG = 'log',
  PULL = 'pull',
  MERGE = 'merge',
  PUSH = 'push',
  RM = 'rm',
  REVERT = 'revert',
  REBASE = 'rebase',
  STATUS = 'status',
  STASH = 'stash',
  TAG = 'tag',
  WORKTREE = 'worktree'
}